Backup register and warehouse DB snapshots to S3
This adds a new module "rds_export_to_s3" that is largely based on
https://github.com/binbashar/terraform-aws-rds-export-to-s3/tree/master

It sets up an S3 bucket for storing database snapshots for a time
(7 days pre-production and 21 days for production).

Lambda functions take care of triggering and monitoring exports of RDS
snapshots to S3.
See the README added as part of this commit for full details of the module.