Prevent Terraform updating Lambdas each TF run
There is a known issue with the terraform-aws-lambda module
that it will re-create new zip files for each run, even if
there is no change to the function.

The comment here suggests setting the "recreate_missing_package"
param to false to avoid this:
https://github.com/terraform-aws-modules/terraform-aws-lambda/issues/396#issuecomment-1612827096

Tweak to the .gitignore to allow these specific zip files to
be committed, so all devs have them in their filesystem.