Pedantically fix files without a terminating line feed (/n)
This used to be a thing that people cared about.

People *should* still care about it. I know we're not using POSIX
any more, but there are still good reasons for ending files with a
line terminator.

See this classic SO thread for some reasons:
https://stackoverflow.com/questions/729692/why-should-text-files-end-with-a-newline

Also kill the final (hopefully!) CRLF Windows-style line breaks from the repo.