EPBR-6353 adds test ecs task stage to fluenbit pipeline
The new stage runs a shell command to start a ECS front end task in intergration and check it's status. If the task stops the stage fails and the pipeline is blocked. This will require the integration ECS front end task to set the fluentbut container to required in the task definition