EPBR-6535 updates how Frontend ecs task are deployed integration
Updates the ECS task's minimum healthy percent to 0. This means that restarting front end service will not create a rolling deployment 'blue/green'. All task will be shut down and new tasks will be started. This is to ensure the smoke test are testing only newly deployed code. Also updated the same task so that the fluentbit container is essential in Integration. This allow us to run more accurate tests on both containers