EPBR-6531 updates the Redis version
Updates Redis to version 7.1. The update required a change to the associated parameter group to the redis7 family. The AWS provided parameter group is the one used as it is not possible to terraform a change to an existing param group. The provided  group has the same 'maxmemory-policy' param as the resource previously created via TF.