EPBR-7195 add schedule task for address base update
Moves the address base task into the schedule app so that it's output is more inline with the other ecr images. Updates the register schedule task module to additionally receive the image name and task arn of the node task. The schedule uses the new task info in it's  configuration