EPBR-7439 adds a custom encryption key
This is the key to be used to re-encrypt the reg api db. The key is not multi region. It is set to rotate every 90 days in Integration and 1 year elsewhere, 90 days being the minimum rotation period allowed. If after that period the key rotation causes any db issues, our smoke test will fail and we will have 9 months to deal with the same issue in prod.