Increase 5xx alarm threshold to 20
We have a lot of alerts in Slack that hit bang on 10
5xx. These are invariably 504/502, we think as scaling
up/down happens, and there's no action we take in response.

By doubling the limit to 20 it should significantly reduce
the noise from these alerts, but is still al low enough
number that we should be alerted swiftly of any genuine
issue with the service.