EPBR-7439 imports encrypted db via new module instance
A new db cluster has been created via the CLI that is a copy of the reg api encrypted with the key in the previous commit. This change adds a new module for the db cluster 'register_api_database-v2'. This module calls the existing aurora-rds module. This been updated to include the naming of the newly created cluster by receiving a suffix var passed to it and create the relevant resources for the existing db (security group, IAM roles etc)