EPBR-7439 fix output for imported db
When importing the db cluster into the state the output file raises an interpolation error. This is because the state does not at that point know about db instances and so cannot create a URL string to connect to them. This changed fixes that issue by adding flow control to the output.