Use cheaper CloudFront distros and remove unused ACM certs
The PriceClass100 covers London, US and Europe which should be plenty
for our users.

Removes Londn-based certs for tech docs and API docs, as these are
behind Cloudfront and use the N.Virginia certificates.

Removes N.virginia certificate for prototypes, which uses the London-
based certificate in front of a load balancer.