Creates an RDS cluster param group for MD5 encryption
Updates the existing param group module so we can create a separate param group for the DWH db cluster that has MD5 encryption as the password param. This change is required when using AWS Glue which uses a JDBC driver which in turn uses MD5 for its password. At present will only be using Glue to connect to the DWH cluster