EPBR-8820: Add TF to import toggles RDS instance
Using a script we can create an encrypted version of the toggles db using the CLI. This code adds a new db cluster and calls the existing rds module. Additional secrets have also been created and is based on the following changes made when applied to the reg-api-db (commit 8435eaa).