Fix `tf-env` just command to work first time
Running `tf-env` failed to work first time because changes to the $AWS_PROFILE environment variable made during the `just set-profile` step are not immediately reflected in the lifetime of the just invocation.

The environment is read into memory at the start of execution of just so updates to the shell environment during the execution don't carry through.  This is why the "tfvars-get" and "tf-init" subsequent to the "set-profile" still picked up the original AWS_PROFILE rather than the updated one.

The change here makes it so the profile can optionally be explicitly passed in to the "tfvars-get" and "tf-init" recipes, so the "tf-env" command can explicitly pass in the updated profile value that can't be read from the environment.