Updates insert glue job
The existing python script fails when run for domestic RR catalog . Now that RR data does not have a unique pk the update fails. As domestic rr data does not change there is no need to update the catalog. The new code only runs an insert if the catalog is an RR