Update justfile to check that BASTION_HOST is running
When we rebuild the bastion host the old terminated instance hangs around for up to 24 hours, and the ID is returned along with the ID of the new running instance.  This causes the just commands that rely on $BASTION_HOST to fail.

This adds an additional filter to the query so that only running instances are returned, of which there should only ever be at most one called "bastion-host".