Update the bastion host operating system to ubuntu 24.04
The bastion host was out of LTS so we've updated it to use the latest LTS version. We've removed local-exec provider block as this is not required to install psql or redis command line as these already exist in ubuntu. Also, removed the ignore changes for the aws ami because this was preventing security updates being updated over time.